Shared Clipboard/URI: Update.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@78581 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/VBox/GuestHost/SharedClipboard-uri.h
@@ -88,7 +88,7 @@ protected:
 protected:
 
     /** Number of references to this instance. */
-    volatile uint16_t            m_cRefs;
+    volatile uint32_t            m_cRefs;
     /** Critical section for serializing access. */
     RTCRITSECT                   m_CritSect;
     /** Open flags. */
